LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/update.https-expected.txt

CONSOLE MESSAGE: Not allowed to follow a redirection while loading https://localhost:9443/service-workers/service-worker/resources/update-worker.py?Key=7cca7f1f-2ac8-4721-86d1-d11813406efe&Mode=redirect
CONSOLE MESSAGE: Cannot load https://localhost:9443/service-workers/service-worker/resources/update-worker.py?Key=7cca7f1f-2ac8-4721-86d1-d11813406efe&Mode=redirect due to access control checks.

PASS update() should succeed when new script is available. 
PASS update() should fail when mime type is invalid. 
FAIL update() should fail when a response for the main script is redirect. assert_throws: function "function () { throw e }" threw object "SecurityError: Script https://localhost:9443/service-workers/service-worker/resources/update-worker.py?Key=7cca7f1f-2ac8-4721-86d1-d11813406efe&Mode=redirect load failed" ("SecurityError") expected object "TypeError" ("TypeError")
PASS update() should fail when a new script contains a syntax error. 
PASS update() should resolve when the install event throws. 
PASS update() should fail when the pending uninstall flag is set. 
PASS update() should succeed when the script shrinks. 
